pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "openpyxl>=3.1",
    "pdfplumber>=0.11",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
